Added recommended repo
Fixing issue #10
