+++ Android/app/src/main/assets/src/language/languages.d.ts
@@ -4,8 +4,8 @@ export declare const languages: import("localized-strings").LocalizedStrings<{
     repository: string;
     repositories: string;
     custom_repository: string;
-    appearance: string;
     accent_color: string;
+    appearance: string;
     language: string;
     dark_theme: string;
     bottom_navigation_text: string;
@@ -42,8 +42,8 @@ export declare const languages: import("localized-strings").LocalizedStrings<{
     repository: string;
     repositories: string;
     custom_repository: string;
-    accent_color: string;
     appearance: string;
+    accent_color: string;
     language: string;
     dark_theme: string;
     bottom_navigation_text: string;
